+++ different-architecture.md
