+++ lib/math.ts
