+++ result.csv
@@ -1,5 +1,5 @@
 Operation,Result
-Addition,165
-Subtraction,-11
-Multiplication,6776
-Division,0.875
+Addition,65
+Subtraction,45
+Multiplication,550
+Division,5.5
