+++ src/mockData.ts
@@ -1,0 +1,4 @@
+export const mockArr = [
+  { id: 1, name: "Nishint", role: "Backend-Developer" },
+  { id: 2, name: "Harsh", role: "Frontend-Developer" },
+];
