+++ tsconfig.json
@@ -1,7 +1,7 @@
 {
   "compilerOptions": {
     "target": "es6",
-    "module": "es2022",
+    "module": "commonjs",
     "moduleResolution": "node",
     "rootDir": "./",
     "outDir": "./dist",
